Add parsing of rill.yaml upfront
