+++ runtime/drivers/github/repo.go
@@ -73,6 +73,9 @@ func (c *connection) ListRecursive(ctx context.Context, glob string, skipDirs bo
 
 		// Track file (p is already relative to the FS root)
 		p = filepath.Join("/", p)
+		if drivers.IsIgnored(p, c.ignorePaths) {
+			return nil
+		}
 		entries = append(entries, drivers.DirEntry{
 			Path:  p,
 			IsDir: d.IsDir(),
