+++ runtime/drivers/repo.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"io"
+	"strings"
 	"time"
 
 	runtimev1 "github.com/rilldata/rill/proto/gen/rill/runtime/v1"
@@ -45,4 +46,32 @@ var ErrFileAlreadyExists = errors.New("file already exists")
 type DirEntry struct {
 	Path  string
 	IsDir bool
+}
+
+// ignoredPaths is a list of paths that are ignored by the parser.
+var ignoredPaths = []string{
+	"/tmp",
+	"/.git",
+	"/node_modules",
+}
+
+// IsIgnored returns true if the path (and any files in nested directories) should be ignored.
+func IsIgnored(path string, ignorePathsConfig []string) bool {
+	for _, dir := range ignoredPaths {
+		if path == dir {
+			return true
+		}
+		if strings.HasPrefix(path, dir) && path[len(dir)] == '/' {
+			return true
+		}
+	}
+	for _, dir := range ignorePathsConfig {
+		if path == dir {
+			return true
+		}
+		if strings.HasPrefix(path, dir) && path[len(dir)] == '/' {
+			return true
+		}
+	}
+	return false
 }
